docker: add dc, dce, replacing exec-postgres, exec-redis, exec-web-rails
To start Docker services:

    dc up

To enter a Bash shell for a generic service (e.g. `web`):

    dce web

To enter a PostgreSQL terminal for a service named `postgres`:

    dce postgres

To enter a Redis terminal for a service named `redis`:

    dce redis

To enter a Bash shell for an overridden service (e.g. `postgres`):

    dce postgres bash

To run some other command for any service (e.g. `redis`):

    dce redis pwd